requirements.txt
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.25
python-dotenv==1.0.0 